# docs/swagger.yaml
basePath: /api/v1
definitions:
  errorHandling.AppError:
    properties:
      code:
        type: integer
      key:
        type: string
      log:
        type: string
      message:
        type: string
    type: object
...
  models.Category:
    properties:
      Image:
        type: string
      Name:
        type: string
      Products:
        items:
          $ref: '#/definitions/models.Product'
        type: array
      ProductsStore:
        items:
          $ref: '#/definitions/models.ProductStore'
        type: array
      Slug:
        type: string
      created_at:
        type: integer
      id:
        $ref: '#/definitions/models.UID'
        type: object
      updated_at:
        type: integer
    type: object
...
  models.Image:
    properties:
      Url:
        type: string
      created_at:
        type: integer
      id:
        $ref: '#/definitions/models.UID'
        type: object
      updated_at:
        type: integer
    type: object
...
  models.Product:
    properties:
      ActiveElement:
        type: string
      Avatar:
        type: string
      Barcode:
        type: string
      BasePrice:
        type: number
      Category:
        items:
          $ref: '#/definitions/models.Category'
        type: array
      Code:
        type: string
      ConditionVoucher:
        type: string
      Content:
        type: string
      Cost:
        type: number
      Description:
        type: string
      Direction:
        type: string
      DoNotUse:
        type: string
      DrugInteractions:
        type: string
      GlobalManufacturerName:
        type: string
      HasPromote:
        type: boolean
      HasPromoteVoucher:
        type: boolean
      Images:
        items:
          $ref: '#/definitions/models.Image'
        type: array
      IndicationsOfTheDrug:
        type: string
      Manufacturer:
        type: string
      Name:
        type: string
      Overdose:
        type: string
      PackagingSize:
        type: string
      Percent:
        type: number
      RegistrationNo:
        type: string
      Status:
        type: string
      Storage:
        type: string
      Tags:
        items:
          $ref: '#/definitions/models.Tag'
        type: array
      Unit:
        type: string
      ValueVoucher:
        type: number
      Variants:
        items:
          $ref: '#/definitions/models.Variant'
        type: array
      Voucher:
        $ref: '#/definitions/models.Voucher'
        type: object
      VoucherId:
        $ref: '#/definitions/models.UID'
        type: object
      created_at:
        type: integer
      id:
        $ref: '#/definitions/models.UID'
        type: object
      updated_at:
        type: integer
    type: object
  models.ProductStore:
    properties:
      ActiveElement:
        type: string
      Avatar:
        type: string
      Barcode:
        type: string
      BasePrice:
        type: number
      Category:
        items:
          $ref: '#/definitions/models.Category'
        type: array
      Code:
        type: string
      Content:
        type: string
      Description:
        type: string
      Direction:
        type: string
      DoNotUse:
        type: string
      DrugInteractions:
        type: string
      GlobalManufacturerName:
        type: string
      Images:
        items:
          $ref: '#/definitions/models.Image'
        type: array
      IndicationsOfTheDrug:
        type: string
      Manufacturer:
        type: string
      Name:
        type: string
      Overdose:
        type: string
      PackagingSize:
        type: string
      RegistrationNo:
        type: string
      Status:
        type: string
      Storage:
        type: string
      Tags:
        items:
          $ref: '#/definitions/models.Tag'
        type: array
      Unit:
        type: string
      Variants:
        items:
          $ref: '#/definitions/models.Variant'
        type: array
      created_at:
        type: integer
      id:
        $ref: '#/definitions/models.UID'
        type: object
      updated_at:
        type: integer
    type: object
...
  models.Tag:
    properties:
      Name:
        type: string
      Slug:
        type: string
      created_at:
        type: integer
      id:
        $ref: '#/definitions/models.UID'
        type: object
      updated_at:
        type: integer
    type: object
  models.UID:
    type: object
...
  models.Variant:
    properties:
      Name:
        type: string
      created_at:
        type: integer
      id:
        $ref: '#/definitions/models.UID'
        type: object
      updated_at:
        type: integer
      variantValue:
        items:
          $ref: '#/definitions/models.VariantValue'
        type: array
    type: object
  models.VariantValue:
    properties:
      convertValue:
        type: number
      operator:
        type: string
      product:
        $ref: '#/definitions/models.Product'
        type: object
      productID:
        type: integer
      variant:
        $ref: '#/definitions/models.Variant'
        type: object
      variantID:
        type: integer
    type: object
  models.Voucher:
    properties:
      Deleted:
        type: boolean
      MaxValue:
        type: number
      Name:
        type: string
      Note:
        type: string
      Type:
        type: string
      Unit:
        type: string
      Value:
        type: number
      created_at:
        type: integer
      id:
        $ref: '#/definitions/models.UID'
        type: object
      updated_at:
        type: integer
    type: object
...
  requests.ChangeStatusProductsRequest:
    properties:
      ProductsId:
        items:
          $ref: '#/definitions/models.UID'
        type: array
      Status:
        example: show/hide/approve/cancel/outofstock
        type: string
    type: object
...
  requests.ProductRequest:
    properties:
      ActiveElement:
        example: example
        type: string
      Avatar:
        example: example
        type: string
      Barcode:
        example: example
        type: string
      BasePrice:
        example: 1
        type: number
      Categories:
        items:
          $ref: '#/definitions/models.UID'
        type: array
      Code:
        example: MD01
        type: string
      Content:
        example: example
        type: string
      Description:
        example: example
        type: string
      Direction:
        example: example
        type: string
      DoNotUse:
        example: example
        type: string
      DrugInteractions:
        example: example
        type: string
      GlobalManufacturerName:
        example: example
        type: string
      IndicationsOfTheDrug:
        example: example
        type: string
      Manufacturer:
        example: abc
        type: string
      Name:
        example: name
        type: string
      Overdose:
        example: example
        type: string
      PackagingSize:
        example: example
        type: string
      RegistrationNo:
        example: example
        type: string
      Status:
        example: show/hide/approve/cancel/outofstock
        type: string
      Storage:
        example: example
        type: string
      Tags:
        items:
          $ref: '#/definitions/models.UID'
        type: array
      Unit:
        example: example
        type: string
      Variants:
        items:
          $ref: '#/definitions/models.UID'
        type: array
    type: object
...
  requests.SearchProductRequest:
    properties:
      AreaId:
        $ref: '#/definitions/models.UID'
        type: object
      Barcode:
        example: example
        type: string
      Category:
        $ref: '#/definitions/models.UID'
        type: object
      Code:
        example: MD01
        type: string
      Name:
        example: name
        type: string
      Status:
        example: show/hide/approve/cancel/outofstock
        type: string
      limit:
        example: 10
        type: integer
      offset:
        example: 0
        type: integer
      sort:
        $ref: '#/definitions/requests.SortOption'
        type: object
      time_from:
        type: integer
      time_to:
        type: integer
    type: object
...
  requests.SearchPureProductRequest:
    properties:
      Barcode:
        example: example
        type: string
      Category:
        $ref: '#/definitions/models.UID'
        type: object
      Code:
        example: MD01
        type: string
      Name:
        example: name
        type: string
      Status:
        example: show/hide/approve/cancel/outofstock
        type: string
      limit:
        example: 10
        type: integer
      offset:
        example: 0
        type: integer
      sort:
        $ref: '#/definitions/requests.SortOption'
        type: object
      time_from:
        type: integer
      time_to:
        type: integer
    type: object
...
  requests.SearchSuggestRequest:
    properties:
      Name:
        example: name
        type: string
    type: object
...
  requests.SortOption:
    properties:
      sort_direction:
        type: string
      sort_field:
        type: string
    type: object
...
  responses.Data:
    properties:
      code:
        type: integer
      key:
        type: string
      message:
        type: string
    type: object
...
  responses.MessageDetail:
    properties:
      listProductChangeStatusFail:
        items:
          type: integer
        type: array
      listProductChangeStatusSuccess:
        items:
          type: integer
        type: array
      listProductNotFound:
        items:
          type: integer
        type: array
    type: object
...
  responses.ProductSearch:
    properties:
      code:
        type: integer
      data:
        items:
          $ref: '#/definitions/models.Product'
        type: array
      message:
        type: string
      total:
        type: integer
    type: object
...
host: v1.api.medilane.vn
info:
  contact:
    name: medilane team
    url: https://www.medilane.vn/
  description: This is openapi for account api.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Medilane account api
  version: "1.0"
paths:
...
  /product:
    post:
      consumes:
      - application/json
      description: Perform create product
      operationId: create-product
      parameters:
      - description: Filter product
        in: body
        name: params
        required: true
        schema:
          $ref: '#/definitions/requests.ProductRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/responses.Data'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errorHandling.AppError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errorHandling.AppError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/errorHandling.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errorHandling.AppError'
      security:
      - BearerAuth: []
      summary: Create product in system
      tags:
      - Product Management
  /product/{id}:
    delete:
      consumes:
      - application/json
      description: Perform delete product
      operationId: delete-product
      parameters:
      - description: id product
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.Data'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errorHandling.AppError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errorHandling.AppError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/errorHandling.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errorHandling.AppError'
      security:
      - BearerAuth: []
      summary: Delete product in system
      tags:
      - Product Management
    get:
      consumes:
      - application/json
      description: Perform get product
      operationId: get-product
      parameters:
      - description: id product
        in: path
        name: id
        required: true
        type: string
      - description: area id
        in: query
        name: area_id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Product'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errorHandling.AppError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errorHandling.AppError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/errorHandling.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errorHandling.AppError'
      security:
      - BearerAuth: []
      summary: Get product in system
      tags:
      - Product Management
    put:
      consumes:
      - application/json
      description: Perform edit product
      operationId: edit-product
      parameters:
      - description: body product
        in: body
        name: params
        required: true
        schema:
          $ref: '#/definitions/requests.ProductRequest'
      - description: id product
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.Data'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errorHandling.AppError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errorHandling.AppError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/errorHandling.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errorHandling.AppError'
      security:
      - BearerAuth: []
      summary: Edit product in system
      tags:
      - Product Management
  /product/{id}/pure:
    get:
      consumes:
      - application/json
      description: Perform get pure product
      operationId: get-pure-product
      parameters:
      - description: id product
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Product'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errorHandling.AppError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errorHandling.AppError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/errorHandling.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errorHandling.AppError'
      security:
      - BearerAuth: []
      summary: Get pure product in system
      tags:
      - Product Management
  /product/find:
    post:
      consumes:
      - application/json
      description: Perform search product
      operationId: search-product
      parameters:
      - description: Filter product
        in: body
        name: params
        required: true
        schema:
          $ref: '#/definitions/requests.SearchProductRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.ProductSearch'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errorHandling.AppError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errorHandling.AppError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/errorHandling.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errorHandling.AppError'
      security:
      - BearerAuth: []
      summary: Search product in system
      tags:
      - Product Management
  /product/pure-search:
    post:
      consumes:
      - application/json
      description: Perform only search product
      operationId: search-pure-product
      parameters:
      - description: Filter product
        in: body
        name: params
        required: true
        schema:
          $ref: '#/definitions/requests.SearchPureProductRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.ProductSearch'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errorHandling.AppError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errorHandling.AppError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/errorHandling.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errorHandling.AppError'
      security:
      - BearerAuth: []
      summary: Search only product in system
      tags:
      - Product Management
  /product/suggest:
    post:
      consumes:
      - application/json
      description: Perform suggest product
      operationId: suggest-product
      parameters:
      - description: Suggest product
        in: body
        name: params
        required: true
        schema:
          $ref: '#/definitions/requests.SearchSuggestRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.ProductSearch'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errorHandling.AppError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errorHandling.AppError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/errorHandling.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errorHandling.AppError'
      security:
      - BearerAuth: []
      summary: Suggest product in system
      tags:
      - Product Management
  /products/status:
    post:
      consumes:
      - application/json
      description: Perform Change status of list product
      operationId: change-status-products
      parameters:
      - description: body change status products
        in: body
        name: params
        required: true
        schema:
          $ref: '#/definitions/requests.ChangeStatusProductsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.MessageDetail'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errorHandling.AppError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errorHandling.AppError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/errorHandling.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errorHandling.AppError'
      security:
      - BearerAuth: []
      summary: Change status of list product in system
      tags:
      - Product Management
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"